// portal-frontend/src/app/interfaces/lease.ts
export interface ILease {
  id?: number;
  tenantId?: number;
  tenantid?: number;
  tenantName?: string;
  propertyId?: number;
  propertyid?: number;
  propertyAddress?: string;
  leaseStart?: string;
  leaseEnd?: string;
  leasetermstart?: string;
  leasetermend?: string;
  paymentStatus?: string;
  paymentStatusId?: number;
  paymentstatusid?: number;
}